# www/api/openapi.yaml
openapi: 3.0.3
info:
  title: RootWarden API
  description: |
    API REST du backend Flask de RootWarden — plateforme de gestion centralisee
    de serveurs Linux (cles SSH, mises a jour, iptables, scan CVE).

    **Authentification** : header `X-API-KEY` obligatoire sur toutes les routes (sauf /test).
    Le frontend PHP utilise `/api_proxy.php` qui injecte la cle automatiquement.
  version: 1.11.0
  contact:
    name: RootWarden
    url: https://github.com/Timikana/rootwarden
servers:
  - url: https://localhost:5000
    description: Backend direct (dev)
  - url: /api_proxy.php
    description: Via proxy PHP (production)
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Erreur description"
...
security:
  - ApiKeyAuth: []
paths:
...
  /server_status:
    post:
      tags: [Monitoring]
      summary: Verifie le statut online/offline d'un serveur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ip]
              properties:
                ip: { type: string, example: "10.0.0.1" }
                port: { type: integer, default: 22 }
      responses:
        '200':
          description: Statut du serveur
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  ip: { type: string }
                  status: { type: string, enum: [online, offline] }
        '400': { description: IP manquante, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  /filter_servers:
    get:
      tags: [Monitoring]
      summary: Filtre les serveurs par environnement, criticite, reseau, tag
      parameters:
        - { name: environment, in: query, schema: { type: string } }
        - { name: criticality, in: query, schema: { type: string } }
        - { name: networkType, in: query, schema: { type: string } }
        - { name: tag, in: query, schema: { type: string } }
      responses:
        '200': { description: Serveurs filtres }
...
  /last_reboot:
    post:
      tags: [Monitoring]
      summary: Date du dernier redemarrage + indicateur reboot_required
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [machine_id]
              properties:
                machine_id: { type: integer }
      responses:
        '200': { description: Date et statut reboot }
...